Use setncattr rather than setattr.
Use the netCDF4.Variable setncattr method to set the attributes of a
variable. With newer versions of the libCDF4 package, using setattr
raised an exception when the attribute name was a reserved Python
attribute. This was first noticed with netCDF4 1.1.5, but not with
version 1.0.8. This should be backwards compatible with older versions
of netCDF4.