Added setup method to the bmi bridge.
Added a new setup method to wrapped BMI components. The new method is
intended to setup a pre-configured set of input files. If the initialize
method is called without arguments, the setup method is first called and
then initialize is called with the input file (provided by setup) as an
argument.